added more scala to the html
Added more scala code to make it easier to edit the html code.
How the data to fill the html is provided as parameters to the template.
That parameter is used to get some values which are used to set
different values in html (if a radio button is checked or not, what is
the value of an input, etc.)