begin adding user preferences + added new classes
Started implementation to get and save user preferences.
User preferences are being saved in objects, but not in the DB yet. I
think it will be too much work to do this with no real benefit right
now. If Aurelian needs this data, the implementation can done.

Added:
-Airline: a class to hold airline data
-AirlinePreferences: a class to hold user list of prefered airlines
-FlightPreferences: a class to store user preferences regarding flights
(night flights and stopover flights)
-RoutePreferences: a class to store user preferences regarding routes
(cheapest, shortest, with most friends)